Separated funder tests utils from actual tests.
