+++ components/funder/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod utils;
+mod tests;
